% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_enrich_result.R
\name{cluster_enrich_result}
\alias{cluster_enrich_result}
\title{Heirarchical clustering of enrichment results based on the similarity of terms within them}
\usage{
cluster_enrich_result(
  enrich_result,
  cluster_threshold = 0.9,
  min_fold_enrichment = 1,
  max_p_adjust = 0.05,
  clust_method = "average",
  min_cluster_size = 2
)
}
\arguments{
\item{enrich_result}{A clusterProfiler categorical enrichment results from enrichGO or similar}

\item{cluster_threshold}{The tree height at which to create clusters. Higher values give more clusters}

\item{min_fold_enrichment}{Fold Enrichment filter for which terms to cluster}

\item{max_p_adjust}{Adjusted P value filter for which terms to cluster}

\item{clust_method}{How cluster agglomeration method to use - see hclust for details}

\item{min_cluster_size}{The smallest cluster term size to report}
}
\value{
A tibble with only a single gene set for each clustered group
}
\description{
Heirarchical clustering of enrichment results based on the similarity of terms within them
}
