% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_volcano_plot.R
\name{cluster_volcano_plot}
\alias{cluster_volcano_plot}
\title{Create a plot of enrichment vs significance for clustered results}
\usage{
cluster_volcano_plot(cluster_result, interactive = FALSE)
}
\arguments{
\item{cluster_result}{An output from cluster_enrich_result}

\item{interactive}{Whether to return a traditional ggplot (FALSE) or interactive plotly (TRUE) plot}
}
\value{
A ggplot or plotly plot
}
\description{
Create a plot of enrichment vs significance for clustered results
}
