% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_enrichment_results.R
\name{compare_enrichment_results}
\alias{compare_enrichment_results}
\title{Test for differentially enriched categories between gene sets}
\usage{
compare_enrichment_results(enrichment_results)
}
\arguments{
\item{enrichment_results}{A list of results from enrichGO or similar}
}
\value{
A tibble of statistical results for differentially enriched categories
}
\description{
Test for differentially enriched categories between gene sets
}
