% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differential_enrichment_heatmap.R
\name{plot_differential_enrichment_heatmap}
\alias{plot_differential_enrichment_heatmap}
\title{Plot the results of differential enrichment as a heatmap}
\usage{
plot_differential_enrichment_heatmap(
  differential_enrichment_hits,
  scale = TRUE,
  log_scale = TRUE
)
}
\arguments{
\item{differential_enrichment_hits}{The output of compare_enrichment_results}

\item{scale}{A TRUE/FALSE value to say whether to scale the heatmap per row}

\item{log_scale}{Whether to log tranform the enrichment values}
}
\description{
Plot the results of differential enrichment as a heatmap
}
