% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Plot a volcano plot of categorical enrichment results}
\usage{
volcano_plot(enrich_result, interactive = FALSE, title = NA)
}
\arguments{
\item{enrich_result}{A clusterProfiler categorical enrichment results from enrichGO or similar}

\item{interactive}{Whether to return a traditional ggplot (FALSE) or interactive plotly (TRUE) plot}

\item{title}{Text for a title to be placed over the plot}
}
\value{
A ggplot or plotly plot
}
\description{
Plot a volcano plot of categorical enrichment results
}
